refactor and got the language toggle to work by removing memoization from AccordionForm component
