+++ components/Forms/ComponentFactory.tsx
@@ -1,29 +1,10 @@
-import { debounce } from 'lodash'
-import type { Instance } from 'mobx-state-tree'
 import { useRouter } from 'next/router'
 import React, { useState, useEffect } from 'react'
-import type { Form } from '../../client-state/models/Form'
-import type { FormField } from '../../client-state/models/FormField'
-import { RootStore } from '../../client-state/store'
 import { WebTranslations } from '../../i18n/web'
-import { Language } from '../../utils/api/definitions/enums'
-import { FieldType } from '../../utils/api/definitions/fields'
 import MainHandler from '../../utils/api/mainHandler'
-import {
-  useMediaQuery,
-  useStore,
-  useTranslation,
-  useStorage,
-  usePersistentState,
-} from '../Hooks'
-import { CurrencyField } from './CurrencyField'
+import { useMediaQuery, useTranslation } from '../Hooks'
 import { FormButtons } from './FormButtons'
-import { NumberField } from './NumberField'
-import { Radio } from './Radio'
-import { FormSelect } from './Select'
-import { TextField } from './TextField'
-import { AccordionForm, Message } from '@dts-stn/decd-design-system'
-import { ObservedAccordionForm } from './ObservedAccordionForm'
+import { AccordionFormContainer } from './AccordionFormContainer'
 
 /**
  * A component that will receive backend props from an API call and render the data as an interactive form.
@@ -32,24 +13,26 @@ import { ObservedAccordionForm } from './ObservedAccordionForm'
 export const ComponentFactory: React.VFC = () => {
   const router = useRouter()
   const locale = router.locale
-  const tsln = useTranslation<WebTranslations>()
   const isMobile = useMediaQuery(992)
 
   const [data, setData] = useState(null)
 
   useEffect(() => {
     if (sessionStorage.getItem('data')) {
-      setData(JSON.parse(sessionStorage.getItem('data')))
+      const data = JSON.parse(sessionStorage.getItem('data'))
+      setData(data)
     } else {
-      const dataObj = new MainHandler({ _language: locale })
+      const dataObj = new MainHandler({ _language: locale }).results
       sessionStorage.setItem('data', JSON.stringify(dataObj))
       setData(dataObj)
     }
   }, [])
 
-  // useEffect(() => {
-  //   setData(JSON.stringify(data))
-  // }, [data])
+  useEffect(() => {
+    const dataObj = new MainHandler({ _language: locale }).results
+    sessionStorage.setItem('data', JSON.stringify(dataObj))
+    setData(dataObj)
+  }, [locale])
 
   // on mobile only, captures enter keypress, does NOT submit form, and blur (hide) keyboard
   useEffect(() => {
@@ -61,80 +44,18 @@ export const ComponentFactory: React.VFC = () => {
     })
   }, [isMobile])
 
-  // if ('error' in data) {
-  //   // typeof data == ResponseError
-  //   // TODO: when error, the form does not update. Repro: set age to 200, change marital from single to married, notice partner questions don't show
-  //   console.log('Data update resulted in error:', data)
-  // }
-
-  // const root: Instance<typeof RootStore> = useStore()
-  // const form: Instance<typeof Form> = root.form
-
-  // if ('fieldData' in data) {
-  //   // typeof data == ResponseSuccess
-  //   form.setupForm(data.fieldData)
-  //   root.setSummary(data.summary)
-  // }
-
-  const keyStepMap = {
-    step1: {
-      title: tsln.category.age,
-      buttonLabel: tsln.category.incomeDetails,
-      keys: ['age'],
-    },
-    step2: {
-      title: tsln.category.incomeDetails,
-      buttonLabel: tsln.category.legalStatus,
-      keys: ['income', 'skipIncome'],
-    },
-    step3: {
-      title: tsln.category.legalStatus,
-      buttonLabel: tsln.category.residence,
-      keys: ['legalStatus'],
-    },
-    step4: {
-      title: tsln.category.residence,
-      buttonLabel: tsln.category.marital,
-      keys: [
-        'livingCountry',
-        'canadaWholeLife',
-        'yearsInCanadaSince18',
-        'everLivedSocialCountry',
-      ],
-    },
-    step5: {
-      title: tsln.category.marital,
-      buttonLabel: tsln.results,
-      keys: [
-        'maritalStatus',
-        'partnerBenefitStatus',
-        'partnerAge',
-        'partnerLivingCountry',
-        'partnerLegalStatus',
-        'partnerCanadaWholeLife',
-        'partnerYearsInCanadaSince18',
-        'partnerIncome',
-      ],
-    },
-  }
-
-  const [cardsValid, setCardsValid] = useState({
-    step1: { isValid: true },
-    step2: { isValid: true },
-    step3: { isValid: true },
-    step4: { isValid: true },
-    step5: { isValid: true },
-  })
-
   const handleOnChange = (step, field, event) => {
-    if (event.target.value === 23) {
-      setCardsValid((currentCardsData) => {
-        const updatedCardsData = { ...currentCardsData }
-        updatedCardsData['step1'].isValid = true
-        return updatedCardsData
-      })
-    }
+    console.log(`step`, step)
+    console.log(`field`, field)
+    console.log(`event`, event)
 
+    // if (event.target.value === 23) {
+    //   setCardsValid((currentCardsData) => {
+    //     const updatedCardsData = { ...currentCardsData }
+    //     updatedCardsData['step1'].isValid = true
+    //     return updatedCardsData
+    //   })
+    // }
     // if (event.target.value === '23') {
     //   console.log('INSIDE SUCCESS')
     //   setCardsValid((currentCardsData) => {
@@ -149,191 +70,14 @@ export const ComponentFactory: React.VFC = () => {
     //     return updatedCardsData
     //   })
     // }
-
-    field.handleChange(event)
+    // field.handleChange(event)
   }
 
-  const generateCards = (formFields) => {
-    const generateChildren = (step, keys) => {
-      const fields = formFields.filter((field) => keys.includes(field.key))
-      const children = fields.map((field) => {
-        return (
-          <div key={field.key}>
-            {field.type === FieldType.NUMBER && (
-              <div className="pb-4">
-                <NumberField
-                  type={field.type}
-                  name={field.key}
-                  label={field.label}
-                  placeholder={field.placeholder ?? ''}
-                  onChange={debounce(
-                    (e) => handleOnChange(step, field, e),
-                    500
-                  )}
-                  value={field.value}
-                  required
-                />
-              </div>
-            )}
-            {field.type == FieldType.CURRENCY && (
-              <div className="pb-4">
-                <CurrencyField
-                  type={field.type}
-                  name={field.key}
-                  label={field.label}
-                  onChange={debounce(
-                    (e) => handleOnChange(step, field, e),
-                    500
-                  )}
-                  placeholder={field.placeholder ?? ''}
-                  value={field.value}
-                  required
-                />
-              </div>
-            )}
-            {field.type == FieldType.STRING && (
-              <div className="pb-4">
-                <TextField
-                  type={field.type}
-                  name={field.key}
-                  label={field.label}
-                  placeholder={field.placeholder ?? ''}
-                  onChange={debounce(field.handleChange, 300)}
-                  value={field.value}
-                  error={field.error}
-                  required
-                />
-              </div>
-            )}
-            {(field.type == FieldType.DROPDOWN ||
-              field.type == FieldType.DROPDOWN_SEARCHABLE) && (
-              <div className="pb-4">
-                <FormSelect
-                  name={field.key}
-                  field={field}
-                  placeholder={getPlaceholderForSelect(field, tsln)}
-                  value={null}
-                />
-              </div>
-            )}
-            {(field.type == FieldType.RADIO ||
-              field.type == FieldType.BOOLEAN) && (
-              <div className="pb-4">
-                <Radio
-                  name={field.key}
-                  checkedValue={field.value}
-                  values={
-                    field.type == 'boolean'
-                      ? [
-                          {
-                            key: 'true',
-                            text: tsln.yes,
-                          },
-                          {
-                            key: 'false',
-                            text: tsln.no,
-                          },
-                        ]
-                      : field.values
-                  }
-                  keyforid={field.key}
-                  label={field.label}
-                  onChange={field.handleChange}
-                  required
-                />
-              </div>
-            )}
-            {field.error && (
-              <div className="mt-6 md:pr-12">
-                <Message
-                  id={field.key}
-                  alert_icon_id={field.key}
-                  alert_icon_alt_text="warning icon"
-                  type="warning"
-                  message_heading={field.error}
-                  message_body={field.error}
-                />
-              </div>
-            )}
-            {field.info && (
-              <div className="mt-6 md:pr-12">
-                <Message
-                  id={field.key}
-                  alert_icon_id={field.key}
-                  alert_icon_alt_text="info icon"
-                  type="info"
-                  message_heading={field.info}
-                  message_body={field.info}
-                />
-              </div>
-            )}
-          </div>
-        )
-      })
-
-      return { [step]: children }
-    }
-
-    const cards = Object.keys(keyStepMap).map((step) => {
-      const cardMeta = keyStepMap[step]
-      const children = generateChildren(step, cardMeta.keys) // ex. ("step1", ["age"])
-
-      return {
-        id: step,
-        title: cardMeta.title,
-        buttonLabel: cardMeta.buttonLabel,
-        children: children[step],
-      }
-    })
-
-    return cards
-  }
-
-  const generateCardsValid = (formFields) => {
-    const cardsValid = {}
-    Object.keys(keyStepMap).forEach((step) => {
-      const stepKeys = keyStepMap[step].keys
-      const fields = formFields.filter((field) => stepKeys.includes(field.key))
-      const isValid = !fields.some((field) => field.error)
-
-      cardsValid[step] = { isValid }
-    })
-
-    return cardsValid
-  }
-
-  const renderAccordionForm = (formFields) => {
-    const cards = generateCards(formFields)
-
-    return (
-      <div className="md:w-2/3">
-        <AccordionForm id="mainForm" cardsState={cardsValid} cards={cards} />
-      </div>
-    )
-  }
-
-  const renderData = () => {
-    console.log(data['results'])
-
-    const results = data['results']
-
-    results.fieldData.forEach((field) => {
-      console.log(field.key)
-      console.log(`field.type`, field.type)
-    })
-  }
-
-  // console.log(`form.fields`, form.fields)
-  // console.log(`data['fieldData]`, data['fieldData'])
   return (
     <>
-      {/* {data && renderAccordionForm(data['fieldData'])} */}
-
-      {/* <ObservedAccordionForm form={form} /> */}
-      {data && renderAccordionForm(data['results']['fieldData'])}
-
-      {/* {data && renderData()} */}
-
+      {data && (
+        <AccordionFormContainer data={data} handleOnChange={handleOnChange} />
+      )}
       <div className="grid grid-cols-1 md:grid-cols-3 md:gap-10 mt-10">
         <div className="col-span-2">
           <FormButtons />
@@ -342,49 +86,3 @@ export const ComponentFactory: React.VFC = () => {
     </>
   )
 }
-
-const getPlaceholderForSelect = (
-  field: Instance<typeof FormField>,
-  tsln: WebTranslations
-) => {
-  let text = tsln.selectText[field.key]
-  return text ? text : tsln.selectText.default
-}
-
-// const sanitizeForm = (data) => {
-//   data.map((fieldData) => {
-//     let placeholder,
-//       defaultValue,
-//       options = undefined
-//     if ('default' in fieldData) defaultValue = fieldData.default
-//     if ('placeholder' in fieldData) placeholder = fieldData.placeholder
-//     if ('values' in fieldData) options = fieldData.values
-
-//     // field does not exist, add it
-//     if (!field) {
-//       self.addField({
-//         key: fieldData.key,
-//         type: fieldData.type,
-//         label: fieldData.label,
-//         category: {
-//           key: fieldData.category.key,
-//           text: fieldData.category.text,
-//         },
-//         order: fieldData.order,
-//         placeholder: placeholder,
-//         default: defaultValue,
-//         options: options,
-//         value: defaultValue ?? null,
-//       })
-//       self.fields.sort((a, b) => a.order - b.order)
-//     }
-//     // field does exist, update if any data has changed
-//     else if (field.label !== fieldData.label) {
-//       console.log('updating field ', fieldData.label)
-//       field.label = fieldData.label
-//       field.category = fieldData.category
-//       field.options = options
-//       field.placeholder = placeholder
-//     }
-//   })
-// }
