+++ package.json
@@ -18,8 +18,9 @@
   },
   "dependencies": {
     "@babel/core": "^7.17.9",
-    "@dts-stn/decd-design-system": "^1.28.0",
+    "@dts-stn/decd-design-system": "^1.35.0",
     "@headlessui/react": "^1.5.0",
+    "@solosphere/decd-design-system": "^1.34.0",
     "@tailwindcss/forms": "^0.5.0",
     "@types/lodash": "^4.14.181",
     "csv-writer": "^1.6.0",
@@ -33,7 +34,7 @@
     "prom-client": "^14.0.1",
     "react": "17.0.2",
     "react-dom": "17.0.2",
-    "react-number-format": "^4.9.1",
+    "react-number-format": "^4.9.3",
     "react-select": "^5.2.2",
     "string-strip-html": "^9.1.7",
     "swagger-ui-react": "^4.6.2",
